Split engine into thud/ package; separate responsibilities
Replace the two flat modules (thud.py, thudclasses.py) with a package
that names each layer by responsibility:

  thud/bitboard.py       — Bitboard (289-bit board mask)
  thud/ply.py            — Ply + NoMoveException + notation helpers
  thud/influence_map.py  — InfluenceMap heuristic grid
  thud/gameboard.py      — Gameboard (rules, move/cap enumeration,
                           win conditions). No AI.
  thud/ai_engine.py      — AIEngine + ai_log (heuristic chooser). No
                           rules.
  thud/mcts.py           — MCTSNode scaffolding (to be removed in
                           Phase 5 along with the broken MCTS driver)
  thud/__init__.py       — re-exports the public names so existing
                           callers keep working

Atomic rename, no shim modules. gui.py now imports from `thud` with
an explicit name list; the previous double `from thudclasses import *`
+ `from thud import *` pair is gone. tests/test_bitboard.py imports
from `thud.bitboard` directly.

While moving the Gameboard code, drop the
`Bitboard.create(shift.value >> d)` workaround in find_setups —
those .create() calls re-wrapped the result of the prior broken
__invert__ that produced out-of-range negative ints; with the new
masking __invert__, plain `shift >> d` / `shift << abs(d)` suffices.
Also drop the matching `int(other_map[i[1]])` cast (Bitboard
__getitem__ already returns int).

console.py is intentionally left out of this commit — it currently
contains a broken MCTS skeleton (Phase 5 restores the documented CLI).

Tests: all 55 Bitboard tests pass; 20-move CPU-vs-CPU smoke game
runs to completion via the new package paths.
